+++ app/src/main/java/com/administrator/yaya/bean/invite/TestObligation.java
@@ -234,7 +234,6 @@ public class TestObligation {
             public static class ParamsBean {
             }
         }
-
         public static class OrderStockListBean {
             /**
              * searchValue : null
@@ -281,7 +280,7 @@ public class TestObligation {
             private String payerName;
             private int commodityAmount;
             private Object commoditySurplusAmount;
-            private int commodityPrice;
+            private Double commodityPrice;
             private int orderStatus;
             private String orderRemark;
             private String orderBuildTime;
@@ -426,11 +425,11 @@ public class TestObligation {
                 this.commoditySurplusAmount = commoditySurplusAmount;
             }
 
-            public int getCommodityPrice() {
+            public Double getCommodityPrice() {
                 return commodityPrice;
             }
 
-            public void setCommodityPrice(int commodityPrice) {
+            public void setCommodityPrice(Double commodityPrice) {
                 this.commodityPrice = commodityPrice;
             }
